+++ demo_query.c
@@ -13,6 +13,7 @@
 #include <time.h>
 
 #include "kmip.h"
+#include "kmip_io.h"
 #include "kmip_bio.h"
 #include "kmip_memset.h"
